Improve emissive texture handling
Set emissive texture color values in emissive factor.
This improves the visual likeness drastically
for models which use emissive textures.
